requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-generativeai==0.3.1
python-dotenv==1.0.0 